#13 move packages // rename some things // properties shifted
