+++ src/test/java/be/mathiasbosman/alphaessenergyidproxy/exporter/energyid/EnergyIdWebhookAdapterImplTest.java
@@ -1,4 +1,4 @@
-package be.mathiasbosman.alphaessenergyidproxy.domain.energyid;
+package be.mathiasbosman.alphaessenergyidproxy.exporter.energyid;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
@@ -6,8 +6,7 @@ import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-import be.mathiasbosman.alphaessenergyidproxy.config.EnergyIdProperties;
-import be.mathiasbosman.alphaessenergyidproxy.domain.energyid.dto.MeterReadingsDto;
+import be.mathiasbosman.alphaessenergyidproxy.exporter.energyid.dto.MeterReadingsDto;
 import java.net.MalformedURLException;
 import java.net.URI;
 import java.net.URISyntaxException;
