+++ src/test/java/be/mathiasbosman/alphaessenergyidproxy/util/CollectionUtilsTest.java
@@ -6,19 +6,19 @@ import java.util.Collection;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 
-class StreamUtilsTest {
+class CollectionUtilsTest {
 
   @Test
   void createBatch() {
     List<String> input = List.of("1", "2", "3", "4", "5", "6", "7", "8", "9", "10");
-    Collection<List<String>> batch = StreamUtils.createBatch(input, 2);
+    Collection<List<String>> batch = CollectionUtils.split(input, 2);
     assertThat(batch).hasSize(5);
   }
 
   @Test
   void createBatchWithZeroLength() {
     List<String> input = List.of("1", "2", "3", "4", "5", "6", "7", "8", "9", "10");
-    Collection<List<String>> batch = StreamUtils.createBatch(input, 0);
+    Collection<List<String>> batch = CollectionUtils.split(input, 0);
     assertThat(batch).hasSize(1);
   }
 }
